Add lower and upper bounds to node
The addition of bounds is to speed up the range query process
and filter as many as possible before going into the nodes.
Bounds are set during bulk loading.
Bounds represent the boundaries of the subtree that is rooted at
the current node. Root node is excluded as it covers the entire
space.

TODO: need to fix the bounds in the insertion